move Counter to components
