+++ src/components/Form/Form.js
@@ -1,7 +1,7 @@
 import PropTypes from 'prop-types';
 import React from 'react';
 import { MAX_LENGTH } from '../../constants';
-import Counter from '../../containers/App/Counter';
+import Counter from '../Counter/Counter';
 import ReplyToTweet from '../ReplyToTweet/ReplyToTweet';
 import Textarea from '../Textarea/Textarea';
 import TweetstormButton from '../TweetstormButton/TweetstormButton';
